# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para manter planetas",
        "title": "Planets API",
        "contact": {
            "name": "Thalyson"
        },
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/": {
            "get": {
                "description": "Retorna links para os controllers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Raiz"
                ],
                "summary": "Raiz",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/planetas": {
            "get": {
                "description": "Busca paginada de planetas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planetas"
                ],
                "summary": "Busca planetas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "buscar por nome do planeta",
                        "name": "busca",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pular X itens no resultado",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limite de itens por pagina",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginaResultado"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Insere um novo planeta",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planetas"
                ],
                "summary": "Inserir Planeta",
                "parameters": [
                    {
                        "description": "Dados de um planeta",
                        "name": "planeta_input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PlanetaInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ResultadoInsert"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/planetas/{id}": {
            "get": {
                "description": "Busca um planeta por ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planetas"
                ],
                "summary": "Busca planeta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id do planeta a ser buscado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Planeta"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um planeta usando seu ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planetas"
                ],
                "summary": "Remover Planeta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do planeta a ser deletado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item removido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.PaginaResultado": {
            "type": "object",
            "properties": {
                "dados": {
                    "type": "object"
                },
                "pagina": {
                    "type": "integer"
                },
                "total_paginas": {
                    "type": "integer"
                }
            }
        },
        "models.Planeta": {
            "type": "object",
            "required": [
                "clima",
                "nome",
                "terreno"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "aparicoes_em_filmes": {
                    "type": "integer"
                },
                "atualizado_em": {
                    "type": "string"
                },
                "clima": {
                    "type": "string"
                },
                "criado_em": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "terreno": {
                    "type": "string"
                }
            }
        },
        "models.PlanetaInput": {
            "type": "object",
            "required": [
                "clima",
                "nome",
                "terreno"
            ],
            "properties": {
                "aparicoes_em_filmes": {
                    "type": "integer"
                },
                "clima": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "terreno": {
                    "type": "string"
                }
            }
        },
        "models.ResultadoInsert": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        }
    }
}